+++ tests/modfiles/fallout3/out/prefsini.js
@@ -1,4 +1,4 @@
-export default const prefsini = [
+export default [
   "[General]",
   "SStartingCell=",
   "SCharGenQuest=0001f388",