+++ tests/utils/fallout4Spec.js
@@ -1,7 +1,10 @@
 import test from "ava";
+import chalk from "chalk";
 import { cleanModFile, analyzeFile, scanModDirectory } from "../../src/lib/utils";
+import ini from "../modfiles/fallout4/out/ini";
+import prefsini from "../modfiles/fallout4/out/prefsini";
 
-test.skip("analyzeFile returns correct fallout4 filenames", async t => {
+test("analyzeFile returns correct fallout4 filenames", async t => {
 	const plugins = cleanModFile({
 		filepath: "../modfiles/fallout4/in/plugins.txt"
 	}).then(m => m.filename);
@@ -13,34 +16,29 @@ test.skip("analyzeFile returns correct fallout4 filenames", async t => {
 	t.is(await modlist, "modlist.txt");
 
 	const fallout4 = cleanModFile({
-		filepath: "../modfiles/fallout4/in/fallout4.ini"
-	}).then(m => m.filename);
-	t.is(await fallout4, "fallout4.ini");
+		filepath: "../modfiles/fallout4/in/Fallout4.ini"
+	}).then(file => file.content)
+	.then(content => {
+		content.forEach((line, index) => {
+			if(line !== ini[index]) {
+				console.log(chalk.red(`${line} !== ${ini[index]}`));
+			}
+		});
+		return content;
+	});
+	t.deepEqual(await fallout4, ini);
 
-	const fallout4Prefs = cleanModFile({
-		filepath: "../modfiles/fallout4/in/fallout4prefs.ini"
-	}).then(m => m.filename);
-	t.is(await fallout4Prefs, "fallout4prefs.ini");
-});
+	const fallout4prefs = cleanModFile({
+		filepath: "../modfiles/fallout4/in/Fallout4Prefs.ini"
+	}).then(file => file.content)
+	.then(content => {
+		content.forEach((line, index) => {
+			if(line !== prefsini[index]) {
+				console.log(chalk.red(`${line} !== ${prefsini[index]}`));
+			}
+		});
+		return content;
+	});
+	t.deepEqual(await fallout4prefs, prefsini);
 
-test.skip("analyzeFile returns correct fallout4 file lengths", async t => {
-	const plugins = cleanModFile({
-		filepath: "../modfiles/fallout4/in/plugins.txt"
-	}).then(m => m.content.length);
-	t.is(await plugins, 4);
-
-	const modlist = cleanModFile({
-		filepath: "../modfiles/fallout4/in/modlist.txt"
-	}).then(m => m.content.length);
-	t.is(await modlist, 4);
-
-	const fallout4 = cleanModFile({
-		filepath: "../modfiles/fallout4/in/fallout4.ini"
-	}).then(m => m.content.length);
-	t.is(await fallout4, 6);
-
-	const fallout4Prefs = cleanModFile({
-		filepath: "../modfiles/fallout4/in/fallout4prefs.ini"
-	}).then(m => m.content.length);
-	t.is(await fallout4Prefs, 9);
 });