+++ tests/utils/skyrimSpec.js
@@ -1,4 +1,5 @@
 import test from "ava";
+import chalk from "chalk";
 import { cleanModFile, scanModDirectory } from "../../src/lib/utils";
 import ini from "../modfiles/skyrim/out/ini";
 import prefsini from "../modfiles/skyrim/out/prefsini";
@@ -20,7 +21,7 @@ test("analyzeFile returns correct content", async t => {
 	.then(content => {
 		content.forEach((line, index) => {
 			if(line !== ini[index]) {
-				console.log(line, ini[index]);
+				console.log(chalk.red(`${line} !== ${ini[index]}`));
 			}
 		});
 		return content;
@@ -29,6 +30,14 @@ test("analyzeFile returns correct content", async t => {
 
 	const skyrimPrefs = cleanModFile({
 		filepath: "../modfiles/skyrim/in/SkyrimPrefs.ini"
-	}).then(file => file.content);
-	t.is(await skyrimPrefs, prefsini);
+	}).then(file => file.content)
+	.then(content => {
+		content.forEach((line, index) => {
+			if(line !== prefsini[index]) {
+				console.log(chalk.red(`${line} !== ${prefsini[index]}`));
+			}
+		});
+		return content;
+	});
+	t.deepEqual(await skyrimPrefs, prefsini);
 });